Refactored to separate integrations
- Transforming the code structure to have one file per integration. The
file encapsulates the fetch logic, the view and the integrations.
- Extracted Trello Notifications and Confluence Inline Tasks to two
separate files.
- Updated the main view to display Trello notifications.
