+++ src/trello_notifications.tsx
@@ -1,13 +1,8 @@
-// Consolidate todos in one single place.
-//
-// - Confluence inline-tasks
-// - Jira assigned issues
-// - Jira mentions
-// - Trello notifications
-// - Google Docs mentions
+import * as React from 'react';
 import * as request from 'request';
 
-interface Todo {
+interface Item {
+  key: string;
   origin: string;
   title: string;
   sourceTitle: string;
@@ -17,54 +12,46 @@ interface Todo {
   due: string;
 }
 
-const ATLASSIAN_USERNAME = process.env.ATLASSIAN_USERNAME;
-const ATLASSIAN_PASSWORD = process.env.ATLASSIAN_PASSWORD;
+interface CardProps {
+  title: string;
+}
+
 const TRELLO_KEY = process.env.TRELLO_KEY;
 const TRELLO_TOKEN = process.env.TRELLO_TOKEN;
 
-// Confluence inline-tasks
-function ConfluenceInlineTasks(): Promise<Array<Todo>> {
-  let todos: Array<Todo> = [];
-  return new Promise<Array<Todo>>((resolve, reject) => {
-    request('https://jobteaser.atlassian.net/wiki/rest/mywork/1/task', {
-      'auth': {
-        'user': ATLASSIAN_USERNAME,
-        'pass': ATLASSIAN_PASSWORD,
-      }}, (error, response, body) => {
-        if (error == null && response && response.statusCode === 200) {
-          let tasks = JSON.parse(body);
-          tasks.forEach( (task: any) => {
-            todos.push({
-              origin: 'Confluence',
-              title: task.title,
-              sourceTitle: task.item.title,
-              sourceUrl: task.item.url,
-              details: '',
-              created: task.created,
-              due: ''
-            });
-          });
-          resolve(todos);
-        } else if (error !== null) {
-          reject(new Error(error));
-        } else {
-          reject(new Error('Non-200 status code: ' + response.statusCode + ' (' + body + ')'));
-        }
-      }
+class Card extends React.Component<CardProps, undefined> {
+  render() {
+    return (
+      <div className='todos__card card'>
+        <header className='card-header'>
+          <p className='card-header-title'>
+            {this.props.title}
+          </p>
+        </header>
+        <div className='card-content'>
+          <div className='content'>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus nec iaculis mauris.
+          </div>
+        </div>
+        <footer className='card-footer'>
+          <a href='#' className='card-footer-item'>Archive</a>
+          <a href='#' className='card-footer-item'>Pin</a>
+        </footer>
+      </div>
     );
-  });
+  }
 }
 
-// Trello notifications
-function TrelloNotifications(): Promise<Array<Todo>> {
-  let todos: Array<Todo> = [];
-  return new Promise<Array<Todo>>((resolve, reject) => {
+function Fetch(): Promise<Array<Item>> {
+  let items: Array<Item> = [];
+  return new Promise<Array<Item>>((resolve, reject) => {
     request('https://api.trello.com/1/members/me/notifications', {
       'qs': {
         'key': TRELLO_KEY,
         'token': TRELLO_TOKEN,
       }}, (error, response, body) => {
         if (error == null && response && response.statusCode === 200) {
+          console.log(body);
           let notifications = JSON.parse(body);
           let cardFullName: string;
           notifications.forEach((notification: any) => {
@@ -72,12 +59,14 @@ function TrelloNotifications(): Promise<Array<Todo>> {
               let
                 data = notification.data,
                 board = data.board,
-                card = data.card;
+                card = data.card,
+                itemKey = 'trello-notification__' + notification.type + '__' + notification.date;
 
               switch (notification.type) {
 
                 case 'makeAdminOfOrganization':
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Made admin of organization `' + data.organization.name,
                     sourceTitle: '',
@@ -89,7 +78,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
                   break;
 
                 case 'cardDueSoon':
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Card due soon',
                     sourceTitle: board.name + ' > ' + card.name,
@@ -101,7 +91,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
                   break;
 
                 case 'commentCard':
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Card `' + board.name + ' > ' + card.name + '` commented',
                     sourceTitle: board.name + ' > ' + card.name,
@@ -113,7 +104,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
                   break;
 
                 case 'createdCard':
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Card `' + board.name + ' > ' + card.name + '` created by ' + notification.memberCreator.fullName,
                     sourceTitle: board.name + ' > ' + card.name,
@@ -141,7 +133,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
                     title = 'N/A';
                     console.log(notification);
                   }
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: title,
                     sourceTitle: board.name + ' > ' + card.name,
@@ -154,7 +147,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
 
                 case 'mentionedOnCard':
                   cardFullName = board.name + ' > ' + card.name;
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Mentioned on card `' + cardFullName + '`',
                     sourceTitle: board.name + ' > ' + card.name,
@@ -167,7 +161,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
 
                 case 'addAttachmentToCard':
                   cardFullName = board.name + ' > ' + card.name;
-                  todos.push({
+                  items.push({
+                    key: itemKey,
                     origin: 'Trello',
                     title: 'Added attachment `' + data.name + '` to card `' + cardFullName + '`',
                     sourceTitle: board.name + ' > ' + card.name,
@@ -183,7 +178,7 @@ function TrelloNotifications(): Promise<Array<Todo>> {
               }
             }
           });
-          resolve(todos);
+          resolve(items);
         } else if (error !== null) {
           reject(new Error(error));
         } else {
@@ -195,7 +190,8 @@ function TrelloNotifications(): Promise<Array<Todo>> {
 }
 
 export {
-  Todo,
-  ConfluenceInlineTasks,
-  TrelloNotifications,
+  Item,
+  CardProps,
+  Card,
+  Fetch,
 };
